Merge to main from Develop

+++ src/main/java/com/my_medi/api/userNotification/controller/UserNotificationApiController.java
@@ -1,5 +1,6 @@
 package com.my_medi.api.userNotification.controller;
 
+import com.my_medi.api.common.service.SseService;
 import com.my_medi.api.userNotification.dto.UserNotificationResponseDto.UserNotificationSimplePageResponse;
 import com.my_medi.api.userNotification.dto.UserNotificationResponseDto.UserNotificationDto;
 import com.my_medi.api.common.dto.ApiResponseDto;
@@ -7,13 +8,17 @@ import com.my_medi.api.userNotification.mapper.UserNotificationConverter;
 import com.my_medi.api.userNotification.service.UserNotificationUseCase;
 import com.my_medi.common.annotation.AuthUser;
 import com.my_medi.domain.notification.entity.UserNotification;
+import com.my_medi.domain.notification.exception.UserNotificationHandler;
 import com.my_medi.domain.notification.service.UserNotificationCommandService;
 import com.my_medi.domain.user.entity.User;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
 import java.util.List;
 
@@ -24,6 +29,7 @@ import java.util.List;
 public class UserNotificationApiController {
     private final UserNotificationUseCase userNotificationUseCase;
     private final UserNotificationCommandService userNotificationCommandService;
+    private final SseService sseService;
 
     @Operation(summary = "사용자의 알림을 pagination 으로 조회합니다.")
     @GetMapping
@@ -48,12 +54,20 @@ public class UserNotificationApiController {
                 .readUserNotification(notificationId));
     }
 
-    //TODO list == null 일때 테스트
     @Operation(summary = "사용자의 알림을 삭제합니다.")
     @DeleteMapping
-    public ApiResponseDto<Void> deleteUserNotifications(@RequestParam List<Long> notificationId) {
-        userNotificationCommandService.removeNotifications(notificationId);
+    public ApiResponseDto<Void> deleteUserNotifications(@RequestParam(required = false) List<Long> notificationId) {
+        if (notificationId == null || notificationId.isEmpty()) {
+            throw UserNotificationHandler.EMPTY_NOTIFICATION_ID_LIST;
+        }
 
+        userNotificationCommandService.removeNotifications(notificationId);
         return ApiResponseDto.onSuccess(null);
+    }
+
+    @Operation(summary = "사용자 알림 실시간 조회를 위해 sse 연결을 합니다.")
+    @GetMapping(value = "/stream", produces = MediaType.TEXT_EVENT_STREAM_VALUE)
+    public SseEmitter linkUserAtSse(@AuthUser User user) {
+        return sseService.connectUser(user.getId());
     }
 }
